requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
feedparser>=6.0.10
pyyaml>=6.0.1
pyarrow>=14.0.1
statsmodels>=0.14.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
